* tp/Texinfo/XS/main/parser_conf.h (PARSER_CONF): use size_t for the descriptor field.
* tp/Texinfo/XS/main/build_perl_info.c (build_perl_*)
(element_to_perl_hash, build_additional_info, store_source_mark_list)
(build_tree_to_build, build_string_list, build_elements_list)
(build_errors, add_formatted_error_messages)
(build_target_elements_list, build_identifiers_target)
(build_internal_xref_list, build_float_types_list)
(build_single_index_data, build_global_info, build_global_commands)
(output_unit_to_perl_hash, build_filenames, build_file_counters)
(build_out_filepaths, build_output_files_unclosed_files)
(build_output_files_opened_files),
tp/Texinfo/XS/main/convert_to_texinfo.c (expand_cmd_args_to_texi)
(convert_to_texinfo_internal), tp/Texinfo/XS/main/convert_to_text.c
(convert_to_text_internal), tp/Texinfo/XS/main/convert_utils.c
(find_innermost_accent_contents, definition_arguments_content)
(definition_category_tree), tp/Texinfo/XS/main/debug.c,
tp/Texinfo/XS/main/errors.c (wipe_error_messages),
tp/Texinfo/XS/main/extra.c (get_associated_info_key)
(lookup_associated_info, lookup_associated_info_by_index),
tp/Texinfo/XS/main/get_perl_info.c (apply_sv_parser_conf)
(find_index_entry_sv), tp/Texinfo/XS/main/node_name_normalization.c
(convert_to_normalized_internal), tp/Texinfo/XS/main/parser_conf.c
(retrieve_parser_conf): use size_t instead of int where the variable
could reach a value outside of the int range and also when the index
is better represented by a size_t.

* tp/Texinfo/XS/main/get_perl_info.c (add_svav_to_string_list): use
SSize_t instead of size_t for consistency.

+++ tp/Texinfo/XS/main/convert_to_text.c
@@ -813,7 +813,7 @@ convert_to_text_internal (const ELEMENT *element, TEXT_OPTIONS *text_options,
       else if ((builtin_command_data[data_cmd].flags & CF_brace)
                && builtin_command_data[data_cmd].data == BRACE_inline)
         {
-          int arg_index = 1;
+          size_t arg_index = 1;
           if (data_cmd == CM_inlineraw)
             text_options->raw_state++;
 
@@ -867,7 +867,7 @@ convert_to_text_internal (const ELEMENT *element, TEXT_OPTIONS *text_options,
         {
           if (element->e.c->args.number > 0)
             {
-              int i;
+              size_t i;
               TEXT args_line;
               text_init (&args_line);
               for (i = 0; i < element->e.c->args.number; i++)
@@ -1048,7 +1048,7 @@ convert_to_text_internal (const ELEMENT *element, TEXT_OPTIONS *text_options,
 
   if (element->e.c->contents.number)
     {
-      int i;
+      size_t i;
       int in_code = 0;
       int in_raw = 0;
       if ((data_cmd
