+++ tp/Texinfo/XS/main/parser_conf.h
@@ -39,14 +39,14 @@ typedef struct PARSER_CONF {
     EXPANDED_FORMAT expanded_formats[7];
     VALUE_LIST values;
 
-    /* descriptor is the index in the array of registered conf */
-    int descriptor;
+    /* descriptor is the index +1 in the array of registered conf */
+    size_t descriptor;
 } PARSER_CONF;
 
 extern PARSER_CONF global_parser_conf;
 
 PARSER_CONF *register_conf (void);
-PARSER_CONF *retrieve_parser_conf (int parser_conf_descriptor);
+PARSER_CONF *retrieve_parser_conf (size_t parser_conf_descriptor);
 
 void clear_parser_conf (PARSER_CONF *parser_conf);
